Extracting AuthProviderInfo, migrating usages
