+++ src/Microsoft.Data.SqlClient/netfx/src/Microsoft.Data.SqlClient.csproj
@@ -151,6 +151,9 @@
     <Compile Include="$(CommonSourceRoot)Interop\Windows\SystemErrors.cs">
       <Link>Interop\SystemErrors.cs</Link>
     </Compile>
+    <Compile Include="$(CommonSourceRoot)\Interop\Windows\Sni\AuthProviderInfo.cs">
+      <Link>Interop\Windows\Sni\AuthProviderInfo.cs</Link>
+    </Compile>
     <Compile Include="$(CommonSourceRoot)\Interop\Windows\Sni\ConsumerInfo.cs">
       <Link>Interop\Windows\Sni\ConsumerInfo.cs</Link>
     </Compile>
