+++ src/Microsoft.Data.SqlClient/netfx/src/Microsoft/Data/Interop/SNINativeMethodWrapper.cs
@@ -23,8 +23,6 @@ namespace Microsoft.Data.SqlClient
         private static readonly System.Runtime.InteropServices.Architecture s_architecture = System.Runtime.InteropServices.RuntimeInformation.ProcessArchitecture;
 
         private const int SniOpenTimeOut = -1; // infinite
-        
-        internal delegate IntPtr SqlClientCertificateDelegate(IntPtr pCallbackContext);
 
         internal const int ConnTerminatedError = 2;
         internal const int InvalidParameterError = 5;
@@ -158,24 +156,6 @@ namespace Microsoft.Data.SqlClient
         }
 
         #region Structs\Enums
-        
-
-        [StructLayout(LayoutKind.Sequential)]
-        internal struct AuthProviderInfo
-        {
-            public uint flags;
-            [MarshalAs(UnmanagedType.Bool)]
-            public bool tlsFirst;
-            public object certContext;
-            [MarshalAs(UnmanagedType.LPWStr)]
-            public string certId;
-            [MarshalAs(UnmanagedType.Bool)]
-            public bool certHash;
-            public object clientCertificateCallbackContext;
-            public SqlClientCertificateDelegate clientCertificateCallback;
-            [MarshalAs(UnmanagedType.LPWStr)]
-            public string serverCertFileName;
-        };
 
         internal enum ConsumerNumber
         {
